requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
openai>=1.0.0
python-dotenv>=1.0.0
plotly>=5.15.0
Pillow>=9.5.0
requests>=2.31.0
reportlab>=4.0.0 